+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Darts is still in an early development phase, and we cannot always guarantee bac
 ## [Unreleased](https://github.com/unit8co/darts/tree/master)
 - Added support for retraining model(s) every `n` iteration and on custom condition in `historical_forecasts` method of `ForecastingModel` abstract class. Addressed issues [#135](https://github.com/unit8co/darts/issues/135) and [#623](https://github.com/unit8co/darts/issues/623) by [Francesco Bruzzesi](https://github.com/fbruzzesi).
 - New LayerNorm alternatives, RMSNorm and LayerNormNoBias [#1113](https://github.com/unit8co/darts/issues/1113) by [Greg DeVos](https://github.com/gdevos010).
+- Overwrite min_train_series_length for Catboost and LightGBM [#1214] (https://github.com/unit8co/darts/issues/1214) by [Anne de Vries](https://github.com/anne-devries).
 
 [Full Changelog](https://github.com/unit8co/darts/compare/0.21.0...master)